+++ YARG.Core/Chart/Notes/ProKeysNote.cs
@@ -8,8 +8,8 @@ namespace YARG.Core.Chart
         public ProKeysNoteFlags ProKeysFlags;
 
         public int Key          { get; }
-        public ulong DisjointMask { get; }
-        public ulong NoteMask     { get; private set; }
+        public int DisjointMask { get; }
+        public int NoteMask     { get; private set; }
 
         public bool IsGlissando => (ProKeysFlags & ProKeysNoteFlags.Glissando) != 0;
 
@@ -62,9 +62,9 @@ namespace YARG.Core.Chart
             return new(this);
         }
 
-        private static ulong GetKeyMask(int key)
+        private static int GetKeyMask(int key)
         {
-            return 1UL << key;
+            return 1 << key;
         }
     }
